Remove parameter group
Not needed (at least for now). We treat all parameters the same:
All can be read and written. It's the responsibility of the EPICS
DB maintainer to ensure he only writes to writeable registers.
Writes will be handled on a by register basis. If the DB maintainer
tries to write to a read-only register the device will reply with an
error, the write will fail and the corresponding record will be
switched to alarm.