Move protocol definition into a separate file
The protocol definition (e.g. structs/objects that define what the
bits 'n' bytes of each package stand for) is by its nature reuseable
(in contrast to the implementation of the EPICS driver which is
EPICS specific). The protocol definition could be reused by other
projects like a device simulator.