Use std::thread instead of libCom threads
The libCom threads are lacking a .join() function making it hard
to shut down threads in a clean way. Unfortunately this is true
for both the C and the C++ interface. This causes problems when
running tools like Valgrind. We use std::thread instead to prevent
these issues. The main drawback of this approach is that our threads
cannot be listed on the IOC shell by epicsThreadShowAll. Portability
shouldn't be affected as our code requires C++11 anyway.