Remove unneeded driver parameters
Remove some parameters from the driver that are actually not required
by the driver. IOC engineers can decide to add a PV for these if they
find them useful. Remove write access to driver parameters to prevent
users from messing with the behavior of the driver. We probably want
to implement IOC shell functions to control the log level instead.